scripts: remove petalinux-export.sh
Either export the setup into a separate yocto setup (to be
constructed and maintained according to the needs), or
convert the existing petalinux project into a merge yocto
project where not needed meta layers can be removed, e.g.
meta-petalinux.

Remove the petalinux-export.sh in favor of the convert script.

Signed-off-by: Lothar Rubusch <lothar.rubusch@enclustra.com>