setup: fix cfg fragments
- fix susbstitution mechanism
- fix cfg fragment printing multiple times the option

The first match was replaced. In most cases this was/is probably ok,
but there might be cases where the specific CONFIG_ option is just a
substring of another pattern and removes this setting, by replacing
its setting. The orig substring might be still around with contrary
setting, etc. With few options, chances are low that such flaws pop
up easily.

The current mech should cover all cases:
- CONFIG is already set: ignore
- CONFIG is disabled, but shall be enabled
- CONFIG is enabled, but shall be disabled
- CONFIG is missing

NB: also special cases, e.g. where CONFIG is a list of arguments
are covered now, such as
    CONFIG_ADD_EXTRA_USERS="root:root;petalinux:petalinux;"

The printer in consequence now is working correctly.

Signed-off-by: Lothar Rubusch <lothar.rubusch@enclustra.com>