zynq 7000: update board patch
This is a workaround for backwards compatibility with modules that do not have
the atsha204a chip and use the ds28 instead.
The order in which the eeproms are probed is changed so that the ds28 is used first.
Otherwise, the atsha204a driver is used first and hangs forever in the wakup function
since older revisions of some modules do not have the atsha204a chip.
Hardware example: ZX5 revision <= 2.0

The original bug report can be found here:
https://github.com/enclustra/Mercury_ZX5_PE1_Reference_Design/issues/1

Instead of adding more patches to the atsha204a driver itself, this adjustment
works around the issue by changing the order of probing in the board patch.