dt: remove status property from all dtsi files
The enclustra layers, specifically the reference design, relies on the device tree generator (dtg)
mechanism which uses the XSA as the input to generate the device trees.
The status property is a leftover from when we used the same device trees
in petalinux and ebe and was necessary for ebe only, because ebe did not use the dtg.