yocto: qspi: handle different qspi layouts
We have modules with 64 Mbyte QSPI flash and module with 128 Mbyte QSPI flash.
We cannot use the same QSPI layout for both since the boot.bin of the modules
with 128 Mbyte QSPI flash is larger than the 64 Mbyte QSPI flash layout boot.bin
partition can hold, mainly due to the larger size of the bitstreams for the
Andromeda modules.
In 2024.1, this was solved using a formfactor config for mercury, mars and
andromeda modules. However, the formfactor alone is not sufficient, since we
have Andromeda modules in the near future (xru30), which will have 64 Mbyte QSPI flash.