dt: fixup: add sata node to pe3
Sata is enabled in the Vivado design, but we need to disable it here
in the device tree as the reference design does not require an M.2 disk
to be present. Therefore, we disable the SATA node in the device tree.
Additionally, a proper setup would require configuration of the clock
generator on the pe3, which is also not done in the reference design and
handled in the i2c application note.